apps/api/requirements.txt
# ─── Web Framework ──────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ─── Database ───────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pgvector==0.3.5

# ─── Cache / Queue ──────────────────────────────────────────
redis[hiredis]==5.2.0
celery[redis]==5.4.0

# ─── Auth ───────────────────────────────────────────────────
clerk-backend-api==1.4.0
python-jose[cryptography]==3.3.0
httpx==0.27.2

# ─── AI / LLM ───────────────────────────────────────────────
anthropic==0.39.0
openai==1.55.3
voyageai==0.3.2
llama-index==0.12.0
llama-index-vector-stores-postgres==0.3.0
langgraph==0.2.53
langchain-anthropic==0.3.0

# ─── Payments ───────────────────────────────────────────────
stripe==11.2.0

# ─── Document / File Processing ─────────────────────────────
python-docx==1.1.2
reportlab==4.2.5
openpyxl==3.1.5
pandas==2.2.3
boto3==1.35.70

# ─── Validation / Config ────────────────────────────────────
pydantic==2.10.2
pydantic-settings==2.6.1
email-validator==2.2.0

# ─── Security ───────────────────────────────────────────────
cryptography==43.0.3
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# ─── Observability ──────────────────────────────────────────
structlog==24.4.0
sentry-sdk[fastapi]==2.19.0

# ─── Testing ────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2
factory-boy==3.3.1
